Create number of islands in a matrix.cpp
The problem finds the number of islands( separated chain of '1's in a binary 2D matrix). It uses the concept of DFS (depth first traversal). 
Note- A 1 is said to be connected if it has another 1 around it (either of the 8 directions).

For example-
1 1 0
0 0 1
1 0 1
Here, two islands will be formed
First island will be formed by elements {A[0][0] ,  A[0][1], A[1][2], A[2][2]}
Second island will be formed by {A[2][0]}.